Fix compiler warning from setCursor
I've been working on a project that uses this library and figured I could give back to the community by fixing this compiler warning. On line 347 the max function is used to set a lower limit, but the input data is of type byte. In the arduino environment byte equates to uint8_t and so can never be less than 0 and thus causes a compiler warning.